[IMP] base_exception: Moved test setup into common
